+++ scripts/protocgen.sh
@@ -2,42 +2,20 @@
 
 set -eo pipefail
 
-protoc_gen_gocosmos() {
-  if ! grep "github.com/gogo/protobuf => github.com/regen-network/protobuf" go.mod &>/dev/null ; then
-    echo -e "\tPlease run this command from somewhere inside the sommelier folder."
-    return 1
-  fi
+# get protoc executions
+go get github.com/regen-network/cosmos-proto/protoc-gen-gocosmos 2>/dev/null
 
-  go get github.com/regen-network/cosmos-proto/protoc-gen-gocosmos 2>/dev/null
-}
-
-protoc_gen_doc() {
-  go get -u github.com/pseudomuto/protoc-gen-doc/cmd/protoc-gen-doc 2>/dev/null
-} 
-
-protoc_gen_gocosmos
-protoc_gen_doc
-
+echo "Generating gogo proto code"
 proto_dirs=$(find ./proto -path -prune -o -name '*.proto' -print0 | xargs -0 -n1 dirname | sort | uniq)
 for dir in $proto_dirs; do
-  buf protoc \
-  -I "proto" \
-  -I "third_party/proto" \
-  --gocosmos_out=plugins=interfacetype+grpc,\
-Mgoogle/protobuf/any.proto=github.com/cosmos/cosmos-sdk/codec/types:. \
-  --grpc-gateway_out=logtostderr=true:. \
-  $(find "${dir}" -maxdepth 1 -name '*.proto')
-
+  for file in $(find "${dir}" -maxdepth 1 -name '*.proto'); do
+    if grep go_package $file &>/dev/null; then
+      buf generate --template proto/buf.gen.gogo.yaml $file
+    fi
+  done
 done
 
-# command to generate docs using protoc-gen-doc
-buf protoc \
--I "proto" \
--I "third_party/proto" \
---doc_out=./docs/api \
---doc_opt=./docs/protodoc-markdown.tmpl,proto-docs.md \
-$(find "$(pwd)/proto" -maxdepth 5 -name '*.proto')
-# go mod tidy
+# TODO: command to generate docs using protoc-gen-doc was deleted here
 
 # move proto files to the right places
 cp -r github.com/evmos/ethermint/* ./